+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Supports `array` in `Tuple`
 
 ## 3.11.0 - 2018-08-17
 - Add `merge` and `mergeMatch` methods to `ITuple`